experiment4/requirements.txt
matplotlib==3.8.2
numpy==1.24.1
opencv_python==4.8.1.78
Pillow==9.3.0
Pillow==10.1.0
torch==2.1.1+cu118
torchvision==0.16.1+cu118
